backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
django-cors-headers==4.3.1
Pillow
scikit-learn
pandas==2.2.2
numpy==1.26.4
tensorflow-macos==2.16.1
djangorestframework-simplejwt==5.3.1